+++ trainitem.h
@@ -2,13 +2,13 @@
 
 class TrainItem{
     private:
-        std::vector<int> inputs;
-        std::vector<int> outputs;
+        std::vector<float> inputs;
+        std::vector<float> outputs;
 
     public:
         TrainItem();
-        void setInOut(std::vector<int> input, std::vector<int> output);
-        std::vector<int>* getInputs();
-        std::vector<int>* getOutputs();
+        void setInOut(std::vector<float> input, std::vector<float> output);
+        std::vector<float>* getInputs();
+        std::vector<float>* getOutputs();
         void print();
 };
